requirements.txt
python-telegram-bot==20.7
solana==0.34.0
solders==0.21.0
base58==2.1.1
requests==2.31.0
groq==0.5.0
python-dotenv==1.0.0
aiohttp==3.9.1
websockets==12.0
cachetools==5.3.2